+++ YandexMobileAdsExample/app/src/main/java/com/yandex/ads/sample/adunits/CustomNativeAdActivity.kt
@@ -30,7 +30,7 @@ import com.yandex.mobile.ads.nativeads.NativeAdViewBinder
 class CustomNativeAdActivity : AppCompatActivity(R.layout.activity_custom_native_ad) {
 
     private val nativeAdView get() = binding.nativeAd.root
-    private val adInfoFragment get() = _adInfoFragment!!
+    private val adInfoFragment get() = requireNotNull(_adInfoFragment)
     private val eventLogger = NativeAdEventLogger()
 
     private var nativeAdLoader: NativeAdLoader? = null