+++ YandexMobileAdsExample/app/src/main/java/com/yandex/ads/sample/appopenad/DefaultProcessLifecycleObserver.kt
@@ -1,0 +1,13 @@
+package com.yandex.ads.sample.appopenad
+
+import androidx.lifecycle.DefaultLifecycleObserver
+import androidx.lifecycle.LifecycleOwner
+
+class DefaultProcessLifecycleObserver(
+    private val onProcessCameForeground: () -> Unit
+) : DefaultLifecycleObserver {
+
+    override fun onStart(owner: LifecycleOwner) {
+        onProcessCameForeground()
+    }
+}