+++ changelogs/mediation/mytarget/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to MyTarget Adapter for Yandex Mobile Ads Mediation will be documented in this file.
 
+## Version 5.18.0.0
+
+#### Updated
+* MyTarget SDK version 5.18.0
+* Yandex Mobile Ads SDK 6.0.0
+* Updated minimum supported Yandex Mobile Ads SDK version to 6.0.0
+
 ## Version 5.17.0.0
 
 #### Updated